requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
opencv-python~=4.11.0.86
flask~=3.1.0
streamlit~=1.42.0
pillow~=11.1.0
requests~=2.32.3
matplotlib~=3.9.4
seaborn
scikit-learn

torch~=2.6.0
torchvision~=0.21.0
tqdm~=4.67.1